content_os/pyproject.toml
[project]
name = "content-compliance-engine"
version = "0.1.0"
description = "Rule-based compliance engine for KO/EN content"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "scikit-learn>=1.0",
    "numpy>=1.21.0",
    "httpx>=0.24.0",
    "beautifulsoup4>=4.12.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
